Share createHeader and createFooter implementation for macOS/tvOS and iOS
