+++ Sources/iOS+tvOS/Extensions/UICollectionView+Extensions.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+extension UICollectionView {
+  static var collectionViewHeaderType: String { return UICollectionElementKindSectionHeader }
+  static var collectionViewFooterType: String { return UICollectionElementKindSectionFooter }
+}
