+++ Sources/macOS/Extensions/NSCollectionView+Extensions.swift
@@ -1,6 +1,9 @@
 import Cocoa
 
 public extension NSCollectionView {
+  static var collectionViewHeaderType: NSCollectionView.SupplementaryElementKind { return .sectionHeader }
+  static var collectionViewFooterType: NSCollectionView.SupplementaryElementKind { return .sectionFooter }
+
   convenience public init(frame: CGRect, collectionViewLayout: CollectionViewFlowLayout) {
     self.init(frame: frame)
     self.collectionViewLayout = collectionViewLayout
